LPD-15281 Inline UADApplicationExportController class into UADExportBackgroundTaskExecutor
